+++ snuba/web/rpc/__init__.py
@@ -11,22 +11,41 @@ from sentry_protos.snuba.v1alpha.endpoint_tags_list_pb2 import (
 )
 
 from snuba.utils.metrics.timer import Timer
-from snuba.web.rpc.span_samples import span_samples_query
-from snuba.web.rpc.timeseries.timeseries import timeseries_query
-from snuba.web.rpc.trace_item_attribute_list import trace_item_attribute_list_query
-from snuba.web.rpc.trace_item_attribute_values import trace_item_attribute_values_query
+from snuba.web.rpc.v1alpha.span_samples import span_samples_query
+from snuba.web.rpc.v1alpha.timeseries.timeseries import timeseries_query
+from snuba.web.rpc.v1alpha.trace_item_attribute_list import (
+    trace_item_attribute_list_query,
+)
+from snuba.web.rpc.v1alpha.trace_item_attribute_values import (
+    trace_item_attribute_values_query,
+)
+
+Version = str
+EndpointName = str
 
 ALL_RPCS: Mapping[
-    str, Tuple[Callable[[Any, Timer], ProtobufMessage], type[ProtobufMessage]]
+    Version,
+    Mapping[
+        EndpointName,
+        Tuple[Callable[[Any, Timer], ProtobufMessage], type[ProtobufMessage]],
+    ],
 ] = {
-    "AggregateBucketRequest": (timeseries_query, AggregateBucketRequest),
-    "SpanSamplesRequest": (span_samples_query, SpanSamplesRequest),
-    "TraceItemAttributesRequest": (
-        trace_item_attribute_list_query,
-        TraceItemAttributesRequest,
-    ),
-    "AttributeValuesRequest": (
-        trace_item_attribute_values_query,
-        AttributeValuesRequest,
-    ),
+    "v1alpha": {
+        "AggregateBucketRequest": (timeseries_query, AggregateBucketRequest),
+        "SpanSamplesRequest": (span_samples_query, SpanSamplesRequest),
+        "TraceItemAttributesRequest": (
+            trace_item_attribute_list_query,
+            TraceItemAttributesRequest,
+        ),
+        "AttributeValuesRequest": (
+            trace_item_attribute_values_query,
+            AttributeValuesRequest,
+        ),
+    }
 }
+
+
+def get_rpc_endpoint(
+    name: str, version: str
+) -> Tuple[Callable[[Any, Timer], ProtobufMessage], type[ProtobufMessage]]:
+    return ALL_RPCS[version][name]
